+++ crate/src/detect/mod.rs
@@ -12,6 +12,7 @@ pub(crate) mod extract;
 pub(crate) mod format;
 pub(crate) mod heuristics;
 pub(crate) mod js;
+pub(crate) mod mask;
 pub(crate) mod redos;
 
 mod position;